(Logger): fix electron import issue on renderer thread
